Conflicts:
	lib/python/Components/AVSwitch.py

+++ lib/python/Components/Wol.py
@@ -1,5 +1,6 @@
 from config import config, ConfigSelection, ConfigNothing
 from Components.SystemInfo import SystemInfo
+from Tools.Directories import fileExists
 from boxbranding import getBoxType
 
 class WOL:
@@ -7,15 +8,21 @@ class WOL:
 		pass
 
 	def setWolState(self, value):
-		print '[WOL] set:',value
-		f = open("/proc/stb/fp/wol", "w")
-		f.write(value)
-		f.close()
+		print '[WakeOnLAN] set:',value
+		if fileExists("/proc/stb/fp/wol"):
+			f = open("/proc/stb/fp/wol", "w")
+			f.write(value)
+			f.close()
+		elif fileExists("/proc/stb/power/wol"):
+			f = open("/proc/stb/power/wol", "w")
+			f.write(value)
+			f.close()
 
 def Init():
-	if SystemInfo["WOL"] and not getBoxType() == 'gbquad':
+	if SystemInfo["WakeOnLAN"] and not getBoxType() == 'gbquad':
 		def setWOLmode(value):
 			iwol.setWolState(config.network.wol.value)
+
 		iwol = WOL()
 		config.network.wol = ConfigSelection([("disable", _("No")), ("enable", _("Yes"))], default = "disable")
 		config.network.wol.addNotifier(setWOLmode, initial_call=True)
